refactor(lock): flock singleton 搬至 lock_posix + shim（#84 PORT-4）
POSIX singleton lock 實作 (fcntl.flock) 搬至新檔 sw_core/lock_posix.py，
原 daemon_lock.py 改為 re-export shim，維持向後相容性。
平台 seam 分檔以支援 Windows lock_win.py 平行實作。
